+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.2
+* added 'toggleSelectable' to 'GroupedScrollViewWithToggle', used to to determine whether the item can be selected
+* added 'onTogglePressed' to 'GroupedToggleController', this callback responds when the item is pressed.
+
 ## 0.2.1
 
 * toggleController is no longer destroyed automatically